% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse_age.R
\name{to_human_age_dd}
\alias{to_human_age_dd}
\alias{to_human_age_dw}
\alias{to_human_age_dm}
\alias{to_human_age_dy}
\alias{to_human_age_wd}
\alias{to_human_age_ww}
\alias{to_human_age_wm}
\alias{to_human_age_wy}
\alias{to_human_age_md}
\alias{to_human_age_mw}
\alias{to_human_age_mm}
\alias{to_human_age_my}
\alias{to_human_age_yd}
\alias{to_human_age_yw}
\alias{to_human_age_ym}
\alias{to_human_age_yy}
\title{Convert mouse age to human age equivalent}
\usage{
to_human_age_dd(mouse_age)

to_human_age_dw(mouse_age)

to_human_age_dm(mouse_age)

to_human_age_dy(mouse_age)

to_human_age_wd(mouse_age)

to_human_age_ww(mouse_age)

to_human_age_wm(mouse_age)

to_human_age_wy(mouse_age)

to_human_age_md(mouse_age)

to_human_age_mw(mouse_age)

to_human_age_mm(mouse_age)

to_human_age_my(mouse_age)

to_human_age_yd(mouse_age)

to_human_age_yw(mouse_age)

to_human_age_ym(mouse_age)

to_human_age_yy(mouse_age)
}
\arguments{
\item{mouse_age}{Mouse age}
}
\value{
Corresponding human age
}
\description{
General functions to convert mouse age to equivalent human age based on
https://doi.org/10.1016/j.lfs.2015.10.025. The function suffixes indicate the
units of the age in input and output, e.g., `to_human_age_my` converts a
mouse's age in months to human age in years. The unit abbreviations are d, w,
m, and y, for day, week (7 d), month (365 / 12 = 30.42 d), and year (365 d),
respectively.
}
\examples{
to_human_age_dy(42)
to_human_age_yy(2)

}
